+++ Bencodex/Types/IValue.cs
@@ -1,7 +1,6 @@
 using System;
 using System.Collections.Generic;
 using System.Diagnostics.Contracts;
-using System.IO;
 
 namespace Bencodex.Types
 {
@@ -40,23 +39,6 @@ namespace Bencodex.Types
         [Obsolete("Deprecated in favour of " + nameof(Inspect) + "() method.")]
         string Inspection { get; }
 
-        /// <summary>Encodes the value into <c cref="byte">Byte</c>
-        /// arrays.</summary>
-        /// <returns><c cref="byte">Byte</c> arrays of Bencodex
-        /// representation of the value.</returns>
-        /// <seealso cref="Codec.Encode(IValue)"/>
-        /// <seealso cref="Codec.Encode(IValue, System.IO.Stream)"/>
-        [Pure]
-        IEnumerable<byte[]> EncodeIntoChunks();
-
-        /// <summary>Writes the encoded value into <paramref name="stream"/>.
-        /// </summary>
-        /// <param name="stream">A stream to write the encoded value.</param>
-        /// <seealso cref="Codec.Encode(IValue)"/>
-        /// <seealso cref="Codec.Encode(IValue, System.IO.Stream)"/>
-        /// <seealso cref="EncodeIntoChunks"/>
-        void EncodeToStream(Stream stream);
-
         /// <summary>
         /// Gets a human-readable representation for debugging.
         /// <para>Unloaded values may be omitted.</para>
@@ -66,6 +48,6 @@ namespace Bencodex.Types
         /// <returns>A human-readable representation for debugging, which looks similar to Python's
         /// literal syntax.  However, if a value is a complex tree and contains any unloaded
         /// subvalues, these are omitted and their fingerprints are shown instead.</returns>
-        public string Inspect(bool loadAll);
+        string Inspect(bool loadAll);
     }
 }